Add TS to sponsor etc
