+++ src/components/sponsor/SponsorHighlight.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import sponsors from "../../assets/sponsors.json";
+import { SponsorSection } from "./SponsorSection";
+
+export const SponsorHighlight = () => {
+  return (
+    <div className="sponsor-content">
+      {sponsors.gold.map((sponsor) => {
+        return <SponsorSection sponsor={sponsor} sponsType="brons" />;
+      })}
+    </div>
+  )
+}
